first step

+++ backend2/src/main/java/com/netcracker/edu/backend2/repository/WalletRepository.java
@@ -3,4 +3,5 @@ package com.netcracker.edu.backend2.repository;
 import com.netcracker.edu.backend2.entities.WalletsEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface WalletRepository extends CrudRepository<WalletsEntity, Long> { }
+public interface WalletRepository extends CrudRepository<WalletsEntity, Integer> {
+}
